move to StreamMetadata.ts
