+++ packages/sdk/src/contracts/searchStreams.ts
@@ -4,6 +4,7 @@ import { StreamFactory } from '../StreamFactory'
 import { ChainPermissions, convertChainPermissionsToStreamPermissions, PUBLIC_PERMISSION_USER_ID, StreamPermission } from '../permission'
 import { filter, map, unique } from '../utils/GeneratorUtils'
 import { StreamQueryResult } from './StreamRegistry'
+import { parseMetadata } from '../StreamMetadata'
 
 export interface SearchStreamsPermissionFilter {
     userId: HexString
@@ -48,7 +49,7 @@ export const searchStreams = (
     logger.debug('Search for streams', { term, permissionFilter })
     return map(
         fetchSearchStreamsResultFromTheGraph(term, permissionFilter, orderBy, theGraphClient),
-        (item: SearchStreamsResultItem) => streamFactory.createStream(toStreamID(item.stream.id), Stream.parseMetadata(item.stream.metadata)),
+        (item: SearchStreamsResultItem) => streamFactory.createStream(toStreamID(item.stream.id), parseMetadata(item.stream.metadata)),
         (err: Error, item: SearchStreamsResultItem) => {
             logger.debug('Omit stream from search result (invalid data)', {
                 streamId: item.stream.id,
